fix: Optimizing massive PII retrieve
