+++ src/main/java/it/gov/pagopa/pu/common/pii/citizen/enums/PersonalDataType.java
@@ -1,0 +1,6 @@
+package it.gov.pagopa.pu.common.pii.citizen.enums;
+
+public enum PersonalDataType {
+    SEND_NOTIFICATION
+}
+
