+++ src/main/java/it/gov/pagopa/pu/common/pii/dto/PIIDTO.java
@@ -1,0 +1,4 @@
+package it.gov.pagopa.pu.common.pii.dto;
+
+public interface PIIDTO {
+}
