--- src/main/java/it/gov/pagopa/pu/send/citizen/enums/PersonalDataType.java
@@ -1,6 +1,0 @@
-package it.gov.pagopa.pu.send.citizen.enums;
-
-public enum PersonalDataType {
-    SEND_NOTIFICATION
-}
-
