--- src/main/java/it/gov/pagopa/pu/send/dto/FullPIIDTO.java
@@ -1,8 +1,0 @@
-package it.gov.pagopa.pu.send.dto;
-
-import it.gov.pagopa.pu.send.model.NoPIIEntity;
-
-public interface FullPIIDTO<E extends NoPIIEntity<P>, P extends PIIDTO> {
-  E getNoPII();
-  void setNoPII(E noPII);
-}
