--- src/main/java/it/gov/pagopa/pu/send/dto/PIIDTO.java
@@ -1,4 +1,0 @@
-package it.gov.pagopa.pu.send.dto;
-
-public interface PIIDTO {
-}
