+++ src/main/java/it/gov/pagopa/pu/send/dto/PuRecipientNoPIIDTO.java
@@ -1,5 +1,6 @@
 package it.gov.pagopa.pu.send.dto;
 
+import it.gov.pagopa.pu.common.pii.dto.PIIDTO;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
