+++ src/main/java/it/gov/pagopa/pu/send/dto/SendNotification.java
@@ -1,6 +1,7 @@
 package it.gov.pagopa.pu.send.dto;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
+import it.gov.pagopa.pu.common.pii.dto.FullEntityPIIDTO;
 import it.gov.pagopa.pu.send.dto.generated.LegalFactDTO;
 import it.gov.pagopa.pu.send.enums.NotificationStatus;
 import it.gov.pagopa.pu.send.model.SendNotificationNoPII;
@@ -14,7 +15,7 @@ import lombok.NoArgsConstructor;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class SendNotification implements FullPIIDTO<SendNotificationNoPII, SendNotificationPIIDTO> {
+public class SendNotification implements FullEntityPIIDTO<SendNotificationNoPII, SendNotificationPIIDTO> {
 
   private String sendNotificationId;
   private Long organizationId;
