--- src/main/java/it/gov/pagopa/pu/send/model/NoPIIEntity.java
@@ -1,8 +1,0 @@
-package it.gov.pagopa.pu.send.model;
-
-import it.gov.pagopa.pu.send.dto.PIIDTO;
-
-public interface NoPIIEntity<P extends PIIDTO> {
-  void setPersonalDataId(Long personalDataId);
-  Long getPersonalDataId();
-}
