+++ src/main/java/it/gov/pagopa/pu/send/repository/SendNotificationPIIRepositoryImpl.java
@@ -1,7 +1,8 @@
 package it.gov.pagopa.pu.send.repository;
 
-import it.gov.pagopa.pu.send.citizen.enums.PersonalDataType;
-import it.gov.pagopa.pu.send.citizen.service.PersonalDataService;
+import it.gov.pagopa.pu.common.pii.citizen.enums.PersonalDataType;
+import it.gov.pagopa.pu.common.pii.citizen.service.PersonalDataService;
+import it.gov.pagopa.pu.common.pii.repository.BasePIIRepository;
 import it.gov.pagopa.pu.send.dto.SendNotification;
 import it.gov.pagopa.pu.send.dto.SendNotificationPIIDTO;
 import it.gov.pagopa.pu.send.mapper.SendNotificationPIIMapper;
@@ -19,23 +20,23 @@ public class SendNotificationPIIRepositoryImpl extends BasePIIRepository<SendNot
   }
 
   @Override
-  void setId(SendNotification fullDTO, String id) { fullDTO.setSendNotificationId(id); }
+  protected void setId(SendNotification fullDTO, String id) { fullDTO.setSendNotificationId(id); }
 
   @Override
-  void setId(SendNotificationNoPII noPii, String id) { noPii.setSendNotificationId(id); }
+  protected void setId(SendNotificationNoPII noPii, String id) { noPii.setSendNotificationId(id); }
 
   @Override
-  String getId(SendNotificationNoPII noPii) {
+  protected String getId(SendNotificationNoPII noPii) {
     return noPii.getSendNotificationId();
   }
 
   @Override
-  Class<SendNotificationPIIDTO> getPIITDTOClass() {
+  protected Class<SendNotificationPIIDTO> getPIITDTOClass() {
     return SendNotificationPIIDTO.class;
   }
 
   @Override
-  PersonalDataType getPIIPersonalDataType() {
+  protected PersonalDataType getPIIPersonalDataType() {
     return PersonalDataType.SEND_NOTIFICATION;
   }
 }
